+++ src/benchmark.jsx
@@ -1,5 +1,5 @@
 import { h } from './utils/streamy-hyperscript';
-import { list } from './utils/streamy-createElement';
+import { list } from './utils/streamy-list';
 import { stream, merge$} from './utils/streamy';
 import { render } from './utils/streamy-render';
